kernel: move userspace sem into kernel/sys
This is a kernel permitive for use with userspace, so move it under
kernel.
