+++ kernel/sys/CMakeLists.txt
@@ -1,0 +1,6 @@
+# SPDX-License-Identifier: Apache-2.0
+# Copyright (c) Zephyr Project Contributors
+
+if(NOT CONFIG_USERSPACE)
+  zephyr_sources(sem.c)
+endif()
