+++ kernel/userspace/sem.c
@@ -7,7 +7,6 @@
 #include <zephyr/sys/sem.h>
 #include <zephyr/internal/syscall_handler.h>
 
-#ifdef CONFIG_USERSPACE
 #define SYS_SEM_MINIMUM      0
 #define SYS_SEM_CONTENDED    (SYS_SEM_MINIMUM - 1)
 
@@ -104,36 +103,3 @@ unsigned int sys_sem_count_get(struct sys_sem *sem)
 
 	return (value > SYS_SEM_MINIMUM) ? value : SYS_SEM_MINIMUM;
 }
-#else
-int sys_sem_init(struct sys_sem *sem, unsigned int initial_count,
-		 unsigned int limit)
-{
-	k_sem_init(&sem->kernel_sem, initial_count, limit);
-
-	return 0;
-}
-
-int sys_sem_give(struct sys_sem *sem)
-{
-	k_sem_give(&sem->kernel_sem);
-
-	return 0;
-}
-
-int sys_sem_take(struct sys_sem *sem, k_timeout_t timeout)
-{
-	int ret_value = 0;
-
-	ret_value = k_sem_take(&sem->kernel_sem, timeout);
-	if ((ret_value == -EAGAIN) || (ret_value == -EBUSY)) {
-		ret_value = -ETIMEDOUT;
-	}
-
-	return ret_value;
-}
-
-unsigned int sys_sem_count_get(struct sys_sem *sem)
-{
-	return k_sem_count_get(&sem->kernel_sem);
-}
-#endif
